requirements.txt
streamlit
google-cloud-bigquery
vertexai
db-dtypes